Welcomed Vulkan and Settings revamp
Part 1 of the GUI, Vulkan with a touch of qt.
The GUI can be invoked by putting in any argument in the terminal.
The Vulkan implementation is crude, only allowing 2d bars.
There are some precompiled Vulkan shaders and source in "/vulkan"
There is no real "GUI" or user interface it's only keybindings (from settings.cfg) for control.
Setting has been revamped into diffrent categories "gui", "terminal" and the "shared" for both.
With this commit keybindings can also be set in the "settings.cfg".

With this commit some known issues:
  - Alpha transparency of background on Wayland (qt fault?)
  - Freezing when window becomes not visible
  - The color settings for the GUI is strange
  - No message when changing settings with keybinings like with terminal version
  - Initial resolution is a hack set to 1000x500
  - Old "settings.cfg" is completely incompatible (expect more changes later)